+++ lib/Models/Terria.ts
@@ -38,6 +38,7 @@ import PickedFeatures, {
   featureBelongsToCatalogItem,
   isProviderCoordsMap
 } from "../Map/PickedFeatures";
+import CatalogMemberMixin from "../ModelMixins/CatalogMemberMixin";
 import GroupMixin from "../ModelMixins/GroupMixin";
 import MappableMixin, { isDataSource } from "../ModelMixins/MappableMixin";
 import ReferenceMixin from "../ModelMixins/ReferenceMixin";
@@ -354,7 +355,11 @@ export default class Terria {
   @observable
   baseMaps: BaseMapViewModel[] = [];
 
+  /** ID of basemap pulled from initData - this wil be used **before** `initBaseMapName` */
   initBaseMapId: string | undefined;
+
+  /** Name of basemap pulled from initData - this is for compatibility with sharelinks which use baseMapName instead of baseMapId. This will be used if `initBaseMapId` can't be resolved */
+  initBaseMapName: string | undefined;
 
   previewBaseMapId: string = "basemap-positron";
 
@@ -673,13 +678,6 @@ export default class Terria {
     }
   }
 
-  setBaseMaps(baseMaps: BaseMapModel[]): void {
-    processBaseMaps(baseMaps, this);
-    if (!this.mainViewer.baseMap) {
-      this.loadPersistedOrInitBaseMap();
-    }
-  }
-
   async loadPersistedOrInitBaseMap() {
     // Set baseMap fallback to first option
     let baseMap = this.baseMaps[0].mappable;
@@ -690,12 +688,16 @@ export default class Terria {
     if (baseMapSearch) {
       baseMap = baseMapSearch.mappable;
     } else {
-      console.error(
-        `Couldn't find a basemap for unique id ${persistedBaseMapId}. Trying to load init base map.`
-      );
-      const baseMapSearch = this.baseMaps.find(
-        baseMap => baseMap.mappable.uniqueId === this.initBaseMapId
-      );
+      // Try to find basemap using initBaseMapId and initBaseMapName
+      const baseMapSearch =
+        this.baseMaps.find(
+          baseMap => baseMap.mappable.uniqueId === this.initBaseMapId
+        ) ??
+        this.baseMaps.find(
+          baseMap =>
+            CatalogMemberMixin.isMixedInto(baseMap.mappable) &&
+            baseMap.mappable.name === this.initBaseMapName
+        );
       if (baseMapSearch) {
         baseMap = baseMapSearch.mappable;
       }
@@ -778,6 +780,10 @@ export default class Terria {
 
     if (this.baseMaps.length === 0) {
       processBaseMaps(defaultBaseMaps(this), this);
+    }
+
+    if (!this.mainViewer.baseMap) {
+      this.loadPersistedOrInitBaseMap();
     }
   }
 
@@ -944,6 +950,10 @@ export default class Terria {
       this.initBaseMapId = initData.baseMapId;
     }
 
+    if (isJsonString(initData.baseMapName)) {
+      this.initBaseMapName = initData.baseMapName;
+    }
+
     if (isJsonString(initData.previewBaseMapId)) {
       this.previewBaseMapId = initData.previewBaseMapId;
     } else if (this.initBaseMapId) {
@@ -955,7 +965,7 @@ export default class Terria {
       Array.isArray(initData.baseMaps) &&
       initData.baseMaps.length > 0
     ) {
-      this.setBaseMaps(<BaseMapModel[]>(<unknown>initData.baseMaps));
+      processBaseMaps(<BaseMapModel[]>(<unknown>initData.baseMaps), this);
     }
 
     if (isJsonObject(initData.homeCamera)) {