+++ lib/Models/Workbench.ts
@@ -2,13 +2,14 @@ import i18next from "i18next";
 import { action, computed, observable } from "mobx";
 import filterOutUndefined from "../Core/filterOutUndefined";
 import TerriaError from "../Core/TerriaError";
-import MappableMixin from "../ModelMixins/MappableMixin";
+import CatalogMemberMixin from "../ModelMixins/CatalogMemberMixin";
+import ChartableMixin from "../ModelMixins/ChartableMixin";
 import GroupMixin from "../ModelMixins/GroupMixin";
+import MappableMixin from "../ModelMixins/MappableMixin";
 import ReferenceMixin from "../ModelMixins/ReferenceMixin";
 import TimeFilterMixin from "../ModelMixins/TimeFilterMixin";
 import CommonStrata from "../Models/CommonStrata";
 import LayerOrderingTraits from "../Traits/LayerOrderingTraits";
-import Chartable from "./Chartable";
 import hasTraits from "./hasTraits";
 import { BaseModel } from "./Model";
 
@@ -161,9 +162,8 @@ export default class Workbench {
   /**
    * Adds or removes a model to/from the workbench. If the model is a reference,
    * it will also be dereferenced. If, after dereferencing, the item turns out not to
-   * be {@link AsyncMappableMixin} or {@link Chartable} but it is a {@link GroupMixin}, it will
+   * be {@link AsyncMappableMixin} or {@link ChartableMixin} but it is a {@link GroupMixin}, it will
    * be removed from the workbench. If it is mappable, `loadMapItems` will be called.
-   * If it is chartable, `loadChartItems` will be called.
    *
    * @param item The item to add to or remove from the workbench.
    */
@@ -184,7 +184,7 @@ export default class Workbench {
           target &&
           GroupMixin.isMixedInto(target) &&
           !MappableMixin.isMixedInto(target) &&
-          !Chartable.is(target)
+          !ChartableMixin.isMixedInto(target)
         ) {
           this.remove(item);
         } else if (target) {
@@ -192,12 +192,10 @@ export default class Workbench {
         }
       }
 
+      if (CatalogMemberMixin.isMixedInto(item)) await item.loadMetadata();
+
       if (MappableMixin.isMixedInto(item)) {
         await item.loadMapItems();
-      }
-
-      if (Chartable.is(item)) {
-        await item.loadChartItems();
       }
     } catch (e) {
       this.remove(item);